Refactor converse module from version 2.26.25 to 2.25.63
