+++ instrumentation/aws-bedrock-runtime-converse-api-2.25.63/src/main/java/llm/converse/models/converse/ConverseModelInvocation.java
