+++ instrumentation/aws-bedrock-runtime-converse-api-2.25.63/src/test/java/llm/converse/events/LlmEventTest.java
