+++ instrumentation/aws-bedrock-runtime-converse-api-2.25.63/src/test/java/software/amazon/awssdk/services/bedrockruntime/AIM_Disabled_ServerSide_BedrockRuntimeClient_InstrumentationTest.java
@@ -17,27 +17,26 @@ import com.newrelic.api.agent.Trace;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import software.amazon.awssdk.services.bedrockruntime.model.ConverseStreamRequest;
-import software.amazon.awssdk.services.bedrockruntime.model.ConverseStreamResponseHandler;
+import software.amazon.awssdk.services.bedrockruntime.model.ConverseRequest;
+import software.amazon.awssdk.services.bedrockruntime.model.ConverseResponse;
 
 import java.util.Collection;
 import java.util.Map;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 
 import static llm.converse.events.LlmEvent.LLM_CHAT_COMPLETION_MESSAGE;
 import static llm.converse.events.LlmEvent.LLM_CHAT_COMPLETION_SUMMARY;
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNull;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
-import static software.amazon.awssdk.services.bedrockruntime.MockConverseRequest.converseStreamRequest;
+import static software.amazon.awssdk.services.bedrockruntime.MockConverseRequest.converseRequest;
 
 @RunWith(InstrumentationTestRunner.class)
-@InstrumentationTestConfig(includePrefixes = { "software.amazon.awssdk.services.bedrockruntime" }, configName = "llm_streaming_disabled.yml")
+@InstrumentationTestConfig(includePrefixes = { "software.amazon.awssdk.services.bedrockruntime" }, configName = "llm_disabled_server_side.yml")
 
-public class AIM_StreamingDisabled_BedrockRuntimeAsyncClient_InstrumentationTest {
-    private static final BedrockRuntimeAsyncClientMock mockBedrockRuntimeAsyncClient = new BedrockRuntimeAsyncClientMock();
+public class AIM_Disabled_ServerSide_BedrockRuntimeClient_InstrumentationTest {
+    private static final BedrockRuntimeClientMock mockBedrockRuntimeClient = new BedrockRuntimeClientMock();
     private final Introspector introspector = InstrumentationTestRunner.getIntrospector();
 
     @Before
@@ -46,51 +45,60 @@ public class AIM_StreamingDisabled_BedrockRuntimeAsyncClient_InstrumentationTest
     }
 
     @Test
-    public void testConverseStreamCompletion() throws ExecutionException, InterruptedException {
+    public void testConverseCompletion() {
         boolean isError = false;
-        Void unused = converseStreamRequestInTransaction(converseStreamRequest(isError));
-        assertNull(unused);
-        assertStreamTransaction();
-        assertSupportabilityMetrics();
+        ConverseResponse converseResponse = converseRequestInTransaction(converseRequest(isError));
+
+        assertNotNull(converseResponse);
+        assertNoLlmTransaction();
+        assertNoLlmSupportabilityMetrics();
+        assertNoLlmEvents();
+        assertTrue(introspector.getErrorEvents().isEmpty());
+    }
+
+    @Test
+    public void testConverseCompletionError() {
+        boolean isError = true;
+        ConverseResponse converseResponse = converseRequestInTransaction(converseRequest(isError));
+
+        assertNotNull(converseResponse);
+        assertNoLlmTransaction();
+        assertNoLlmSupportabilityMetrics();
         assertNoLlmEvents();
         assertTrue(introspector.getErrorEvents().isEmpty());
     }
 
     @Trace(dispatcher = true)
-    private Void converseStreamRequestInTransaction(ConverseStreamRequest converseStreamRequest) throws ExecutionException, InterruptedException {
+    private ConverseResponse converseRequestInTransaction(ConverseRequest converseRequest) {
         addCustomParameters();
-
-        ConverseStreamResponseHandler converseStreamResponseHandler = ConverseStreamResponseHandler.builder().subscriber(event -> { }).build();
-        CompletableFuture<Void> converseResponseFuture = mockBedrockRuntimeAsyncClient.converseStream(converseStreamRequest, converseStreamResponseHandler);
-        return converseResponseFuture.get();
+        return mockBedrockRuntimeClient.converse(converseRequest);
     }
 
     private void addCustomParameters() {
-        NewRelic.addCustomParameter("llm.conversation_id", "conversation-id-value");
-        NewRelic.addCustomParameter("llm.testPrefix", "testPrefix");
-        NewRelic.addCustomParameter("test", "test");
+        NewRelic.addCustomParameter("llm.conversation_id", "conversation-id-value"); // Will be added to LLM events
+        NewRelic.addCustomParameter("llm.testPrefix", "testPrefix"); // Will be added to LLM events
+        NewRelic.addCustomParameter("test", "test"); // Will NOT be added to LLM events
     }
 
-    private void assertStreamTransaction() {
+    private void assertNoLlmTransaction() {
         assertEquals(1, introspector.getFinishedTransactionCount(TimeUnit.SECONDS.toMillis(2)));
         Collection<String> transactionNames = introspector.getTransactionNames();
         String transactionName = transactionNames.iterator().next();
         Map<String, TracedMetricData> metrics = introspector.getMetricsForTransaction(transactionName);
-
-        assertTrue(metrics.containsKey("Llm/completion/Bedrock/converseStream"));
-        assertEquals(1, metrics.get("Llm/completion/Bedrock/converseStream").getCallCount());
+        assertFalse(metrics.containsKey("Llm/completion/Bedrock/converse"));
     }
 
-    private void assertSupportabilityMetrics() {
+    private void assertNoLlmSupportabilityMetrics() {
         Map<String, TracedMetricData> unscopedMetrics = introspector.getUnscopedMetrics();
-        assertTrue(unscopedMetrics.containsKey("Supportability/Java/ML/Bedrock/2.26.25"));
-        assertTrue(unscopedMetrics.containsKey("Supportability/Java/ML/Streaming/Disabled"));
+        assertFalse(unscopedMetrics.containsKey("Supportability/Java/ML/Bedrock/2.25.63"));
     }
 
     private void assertNoLlmEvents() {
+        // LlmChatCompletionMessage events
         Collection<Event> llmChatCompletionMessageEvents = introspector.getCustomEvents(LLM_CHAT_COMPLETION_MESSAGE);
         assertEquals(0, llmChatCompletionMessageEvents.size());
 
+        // LlmCompletionSummary events
         Collection<Event> llmCompletionSummaryEvents = introspector.getCustomEvents(LLM_CHAT_COMPLETION_SUMMARY);
         assertEquals(0, llmCompletionSummaryEvents.size());
     }
