+++ instrumentation/aws-bedrock-runtime-converse-api-2.25.63/src/test/java/software/amazon/awssdk/services/bedrockruntime/AIM_StreamingDisabled_BedrockRuntimeAsyncClient_InstrumentationTest.java
@@ -17,26 +17,27 @@ import com.newrelic.api.agent.Trace;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import software.amazon.awssdk.services.bedrockruntime.model.ConverseRequest;
-import software.amazon.awssdk.services.bedrockruntime.model.ConverseResponse;
+import software.amazon.awssdk.services.bedrockruntime.model.ConverseStreamRequest;
+import software.amazon.awssdk.services.bedrockruntime.model.ConverseStreamResponseHandler;
 
 import java.util.Collection;
 import java.util.Map;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 
 import static llm.converse.events.LlmEvent.LLM_CHAT_COMPLETION_MESSAGE;
 import static llm.converse.events.LlmEvent.LLM_CHAT_COMPLETION_SUMMARY;
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
-import static software.amazon.awssdk.services.bedrockruntime.MockConverseRequest.converseRequest;
+import static software.amazon.awssdk.services.bedrockruntime.MockConverseRequest.converseStreamRequest;
 
 @RunWith(InstrumentationTestRunner.class)
-@InstrumentationTestConfig(includePrefixes = { "software.amazon.awssdk.services.bedrockruntime" }, configName = "llm_disabled_server_side.yml")
+@InstrumentationTestConfig(includePrefixes = { "software.amazon.awssdk.services.bedrockruntime" }, configName = "llm_streaming_disabled.yml")
 
-public class AIM_Disabled_ServerSide_BedrockRuntimeClient_InstrumentationTest {
-    private static final BedrockRuntimeClientMock mockBedrockRuntimeClient = new BedrockRuntimeClientMock();
+public class AIM_StreamingDisabled_BedrockRuntimeAsyncClient_InstrumentationTest {
+    private static final BedrockRuntimeAsyncClientMock mockBedrockRuntimeAsyncClient = new BedrockRuntimeAsyncClientMock();
     private final Introspector introspector = InstrumentationTestRunner.getIntrospector();
 
     @Before
@@ -45,60 +46,51 @@ public class AIM_Disabled_ServerSide_BedrockRuntimeClient_InstrumentationTest {
     }
 
     @Test
-    public void testConverseCompletion() {
+    public void testConverseStreamCompletion() throws ExecutionException, InterruptedException {
         boolean isError = false;
-        ConverseResponse converseResponse = converseRequestInTransaction(converseRequest(isError));
-
-        assertNotNull(converseResponse);
-        assertNoLlmTransaction();
-        assertNoLlmSupportabilityMetrics();
-        assertNoLlmEvents();
-        assertTrue(introspector.getErrorEvents().isEmpty());
-    }
-
-    @Test
-    public void testConverseCompletionError() {
-        boolean isError = true;
-        ConverseResponse converseResponse = converseRequestInTransaction(converseRequest(isError));
-
-        assertNotNull(converseResponse);
-        assertNoLlmTransaction();
-        assertNoLlmSupportabilityMetrics();
+        Void unused = converseStreamRequestInTransaction(converseStreamRequest(isError));
+        assertNull(unused);
+        assertStreamTransaction();
+        assertSupportabilityMetrics();
         assertNoLlmEvents();
         assertTrue(introspector.getErrorEvents().isEmpty());
     }
 
     @Trace(dispatcher = true)
-    private ConverseResponse converseRequestInTransaction(ConverseRequest converseRequest) {
+    private Void converseStreamRequestInTransaction(ConverseStreamRequest converseStreamRequest) throws ExecutionException, InterruptedException {
         addCustomParameters();
-        return mockBedrockRuntimeClient.converse(converseRequest);
+
+        ConverseStreamResponseHandler converseStreamResponseHandler = ConverseStreamResponseHandler.builder().subscriber(event -> { }).build();
+        CompletableFuture<Void> converseResponseFuture = mockBedrockRuntimeAsyncClient.converseStream(converseStreamRequest, converseStreamResponseHandler);
+        return converseResponseFuture.get();
     }
 
     private void addCustomParameters() {
-        NewRelic.addCustomParameter("llm.conversation_id", "conversation-id-value"); // Will be added to LLM events
-        NewRelic.addCustomParameter("llm.testPrefix", "testPrefix"); // Will be added to LLM events
-        NewRelic.addCustomParameter("test", "test"); // Will NOT be added to LLM events
+        NewRelic.addCustomParameter("llm.conversation_id", "conversation-id-value");
+        NewRelic.addCustomParameter("llm.testPrefix", "testPrefix");
+        NewRelic.addCustomParameter("test", "test");
     }
 
-    private void assertNoLlmTransaction() {
+    private void assertStreamTransaction() {
         assertEquals(1, introspector.getFinishedTransactionCount(TimeUnit.SECONDS.toMillis(2)));
         Collection<String> transactionNames = introspector.getTransactionNames();
         String transactionName = transactionNames.iterator().next();
         Map<String, TracedMetricData> metrics = introspector.getMetricsForTransaction(transactionName);
-        assertFalse(metrics.containsKey("Llm/completion/Bedrock/converse"));
+
+        assertTrue(metrics.containsKey("Llm/completion/Bedrock/converseStream"));
+        assertEquals(1, metrics.get("Llm/completion/Bedrock/converseStream").getCallCount());
     }
 
-    private void assertNoLlmSupportabilityMetrics() {
+    private void assertSupportabilityMetrics() {
         Map<String, TracedMetricData> unscopedMetrics = introspector.getUnscopedMetrics();
-        assertFalse(unscopedMetrics.containsKey("Supportability/Java/ML/Bedrock/2.26.25"));
+        assertTrue(unscopedMetrics.containsKey("Supportability/Java/ML/Bedrock/2.25.63"));
+        assertTrue(unscopedMetrics.containsKey("Supportability/Java/ML/Streaming/Disabled"));
     }
 
     private void assertNoLlmEvents() {
-        // LlmChatCompletionMessage events
         Collection<Event> llmChatCompletionMessageEvents = introspector.getCustomEvents(LLM_CHAT_COMPLETION_MESSAGE);
         assertEquals(0, llmChatCompletionMessageEvents.size());
 
-        // LlmCompletionSummary events
         Collection<Event> llmCompletionSummaryEvents = introspector.getCustomEvents(LLM_CHAT_COMPLETION_SUMMARY);
         assertEquals(0, llmCompletionSummaryEvents.size());
     }
