+++ instrumentation/aws-bedrock-runtime-converse-api-2.25.63/src/test/java/software/amazon/awssdk/services/bedrockruntime/BedrockRuntimeClientMock.java
