+++ instrumentation/aws-bedrock-runtime-converse-api-2.25.63/src/test/resources/llm_streaming_disabled.yml
