+++ settings.gradle
@@ -70,7 +70,7 @@ include 'instrumentation:activemq-client-5.8.0'
 include 'instrumentation:anorm-2.3'
 include 'instrumentation:anorm-2.4'
 include 'instrumentation:aws-bedrock-runtime-2.20'
-include 'instrumentation:aws-bedrock-runtime-converse-api-2.26.25'
+include 'instrumentation:aws-bedrock-runtime-converse-api-2.25.63'
 include 'instrumentation:aws-java-sdk-dynamodb-1.11.106'
 include 'instrumentation:aws-java-sdk-dynamodb-2.15.34'
 include 'instrumentation:aws-java-sdk-dax-2.0.0'
